requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==20.3.0
autopep8==1.6.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
billiard==3.6.4.0
black==22.6.0
blinker==1.4
Brlapi==0.8.2
cached-property==1.5.2
celery==5.2.7
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmdtest==0.32+git
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
Deprecated==1.2.13
distro==1.5.0
distro-info==1.0
Django==3.2
djangorestframework==3.13.1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.17
fasteners==0.14.1
future==0.18.2
greenlet==0.4.17
gyp==0.1
html5lib==1.1
httplib2==0.18.1
idna==2.10
importlib-metadata==1.6.0
invoke==1.4.1
jsonschema==3.2.0
kombu==5.2.4
lockfile==0.12.2
louis==3.16.0
lxml==4.6.3
Mako==1.1.3
Markdown==3.3.4
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.0
mypy-extensions==0.4.3
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
packaging==21.3
paramiko==2.7.2
pathspec==0.9.0
pexpect==4.8.0
Pillow==8.1.2
platformdirs==2.5.2
prompt-toolkit==3.0.30
psutil==5.8.0
psycopg2-binary==2.9.3
pycairo==1.16.2
pycodestyle==2.8.0
pycups==2.0.1
pycurl==7.43.0.6
Pygments==2.7.1
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==1.7.1
PyNaCl==1.4.0
pynvim==0.4.2
pyparsing==3.0.9
pyrsistent==0.15.5
PySimpleSOAP==1.16.2
pysmbc==1.0.23
python-apt==2.2.1
python-debian==0.1.39
python-debianbts==3.1.0
python-pam==1.8.4
python-xapp==2.0.2
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.3.1
redis==4.3.4
reportbug==7.10.3+deb11u1
requests==2.25.1
scour==0.38.2
setproctitle==1.2.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
texttable==1.6.3
tinycss==0.4
tinycss2==1.0.2
toml==0.10.2
tomli==2.0.1
ttystatus==0.38
typing-extensions==4.3.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
wrapt==1.14.1
xdg==5
zipp==1.0.0
